requirements.txt
flask=3.0.3
waitress=3.0.0
mysql-connector-python=9.0.0
jsonify=0.5